test(otel): move metrics coverage to internal package
