+++ middleware/otel/middleware_test.go
@@ -11,8 +11,6 @@ import (
 	gitlab "gitlab.com/gitlab-org/api/client-go/v2"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
-	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
-	"go.opentelemetry.io/otel/sdk/metric/metricdata"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	"go.opentelemetry.io/otel/sdk/trace/tracetest"
 	"go.opentelemetry.io/otel/trace"
@@ -90,45 +88,6 @@ func TestMiddlewareSkipsNilOption(t *testing.T) {
 	assert.Len(t, recorder.Ended(), 1)
 }
 
-func TestMiddlewareRecordsMetrics(t *testing.T) {
-	reader := sdkmetric.NewManualReader()
-	provider := sdkmetric.NewMeterProvider(sdkmetric.WithReader(reader))
-	event := &gitlab.PushEvent{
-		ObjectKind: "push",
-		EventName:  "push",
-		ProjectID:  123,
-		Ref:        "refs/heads/main",
-		Project: gitlab.PushEventProject{
-			PathWithNamespace: "group/project",
-		},
-	}
-
-	handler := otel.Middleware(otel.WithMeterProvider(provider))(func(context.Context, any) error {
-		return nil
-	})
-
-	err := handler(context.Background(), event)
-	require.NoError(t, err)
-
-	rm := collectMetrics(t, reader)
-	events := metricByName(t, rm, "gitlab.webhook.events")
-	eventSum, ok := events.Data.(metricdata.Sum[int64])
-	require.True(t, ok)
-	require.Len(t, eventSum.DataPoints, 1)
-	assert.Equal(t, int64(1), eventSum.DataPoints[0].Value)
-	assertAttrString(t, eventSum.DataPoints[0].Attributes, "gitlab.webhook.event_type", "push")
-	assertAttrString(t, eventSum.DataPoints[0].Attributes, "gitlab.webhook.object_kind", "push")
-	assertAttrString(t, eventSum.DataPoints[0].Attributes, "gitlab.webhook.result", "success")
-	assertNoAttr(t, eventSum.DataPoints[0].Attributes, "gitlab.project.path")
-	assertNoAttr(t, eventSum.DataPoints[0].Attributes, "gitlab.ref")
-
-	duration := metricByName(t, rm, "gitlab.webhook.event.duration")
-	histogram, ok := duration.Data.(metricdata.Histogram[float64])
-	require.True(t, ok)
-	require.Len(t, histogram.DataPoints, 1)
-	assert.Equal(t, uint64(1), histogram.DataPoints[0].Count)
-}
-
 func attrString(attrs []attribute.KeyValue, key string) string {
 	for _, attr := range attrs {
 		if string(attr.Key) == key {
@@ -147,43 +106,4 @@ func attrInt64(attrs []attribute.KeyValue, key string) int64 {
 	}
 
 	return 0
-}
-
-func collectMetrics(t *testing.T, reader *sdkmetric.ManualReader) metricdata.ResourceMetrics {
-	t.Helper()
-
-	var rm metricdata.ResourceMetrics
-	require.NoError(t, reader.Collect(t.Context(), &rm))
-
-	return rm
-}
-
-func metricByName(t *testing.T, rm metricdata.ResourceMetrics, name string) metricdata.Metrics {
-	t.Helper()
-
-	for _, scope := range rm.ScopeMetrics {
-		for _, metric := range scope.Metrics {
-			if metric.Name == name {
-				return metric
-			}
-		}
-	}
-
-	require.Failf(t, "metric not found", "metric %q was not collected", name)
-	return metricdata.Metrics{}
-}
-
-func assertAttrString(t *testing.T, attrs attribute.Set, key, expected string) {
-	t.Helper()
-
-	value, ok := attrs.Value(attribute.Key(key))
-	require.Truef(t, ok, "attribute %q was not found", key)
-	assert.Equal(t, expected, value.AsString())
-}
-
-func assertNoAttr(t *testing.T, attrs attribute.Set, key string) {
-	t.Helper()
-
-	_, ok := attrs.Value(attribute.Key(key))
-	assert.Falsef(t, ok, "attribute %q should not be recorded", key)
 }
